Split code into separate modules, commented it all, made py 3 compatible, replaced tabs with spaces, added readme, licence and todo list
