+++ todo.md
@@ -1,0 +1,3 @@
+* Test with Windows telnet
+* Test server on Windows
+* Test on mac
